refactor(comment): Moved persistense directory into infrastructure
